+++ examples/basic/src/VideoPlayer.tsx
@@ -34,7 +34,8 @@ import Video, {
   SelectedTrack,
   DRMType,
   OnTextTrackDataChangedData,
-  SelectedTrackType,
+  TextTrackType,
+  ISO639_1,
 } from 'react-native-video';
 import ToggleControl from './ToggleControl';
 import MultiValueControl, {
@@ -126,6 +127,18 @@ class VideoPlayer extends Component {
     {
       description: 'sintel with subtitles',
       uri: 'https://bitmovin-a.akamaihd.net/content/sintel/hls/playlist.m3u8',
+    },
+    {
+      description: 'sintel with sideLoaded subtitles',
+      uri: 'https://demo.unified-streaming.com/k8s/features/stable/video/tears-of-steel/tears-of-steel.ism/.m3u8', // this is sample video, my actual video file is MP4
+      textTracks: [
+        {
+          title: 'test',
+          language: 'en' as ISO639_1,
+          type: TextTrackType.VTT,
+          uri: 'https://bitdash-a.akamaihd.net/content/sintel/subtitles/subtitles_en.vtt',
+        },
+      ],
     },
   ];
 
@@ -225,15 +238,17 @@ class VideoPlayer extends Component {
     const selectedTrack = data.audioTracks?.find((x: AudioTrack) => {
       return x.selected;
     });
-    this.setState({
-      audioTracks: data.audioTracks,
-    });
     if (selectedTrack?.language) {
       this.setState({
+        audioTracks: data.audioTracks,
         selectedAudioTrack: {
           type: 'language',
           value: selectedTrack?.language,
         },
+      });
+    } else {
+      this.setState({
+        audioTracks: data.audioTracks,
       });
     }
   };
@@ -243,16 +258,17 @@ class VideoPlayer extends Component {
       return x?.selected;
     });
 
-    this.setState({
-      textTracks: data.textTracks,
-    });
     if (selectedTrack?.language) {
       this.setState({
-        textTracks: data,
+        textTracks: data.textTracks,
         selectedTextTrack: {
           type: 'language',
           value: selectedTrack?.language,
         },
+      });
+    } else {
+      this.setState({
+        textTracks: data.textTracks,
       });
     }
   };
@@ -638,11 +654,13 @@ class VideoPlayer extends Component {
                 />
               </View>
               <View style={styles.generalControls}>
+                {/* shall be replaced by slider */}
                 <MultiValueControl
-                  values={[0.25, 0.5, 1.0, 1.5, 2.0]}
+                  values={[0, 0.25, 0.5, 1.0, 1.5, 2.0]}
                   onPress={this.onRateSelected}
                   selected={this.state.rate}
                 />
+                {/* shall be replaced by slider */}
                 <MultiValueControl
                   values={[0.5, 1, 1.5]}
                   onPress={this.onVolumeSelected}
@@ -693,6 +711,9 @@ class VideoPlayer extends Component {
                       });
                     }}>
                     {this.state.audioTracks.map(track => {
+                      if (!track) {
+                        return;
+                      }
                       return (
                         <Picker.Item
                           label={track.language}
@@ -720,13 +741,18 @@ class VideoPlayer extends Component {
                       });
                     }}>
                     <Picker.Item label={'none'} value={'none'} key={'none'} />
-                    {this.state.textTracks.map(track => (
-                      <Picker.Item
-                        label={track.language}
-                        value={track.language}
-                        key={track.language}
-                      />
-                    ))}
+                    {this.state.textTracks.map(track => {
+                      if (!track) {
+                        return;
+                      }
+                      return (
+                        <Picker.Item
+                          label={track.language}
+                          value={track.language}
+                          key={track.language}
+                        />
+                      );
+                    })}
                   </Picker>
                 )}
               </View>
@@ -749,6 +775,7 @@ class VideoPlayer extends Component {
             this.video = ref;
           }}
           source={this.srcList[this.state.srcListId]}
+          textTracks={this.srcList[this.state.srcListId]?.textTracks}
           adTagUrl={this.srcList[this.state.srcListId]?.adTagUrl}
           drm={this.srcList[this.state.srcListId]?.drm}
           style={viewStyle}
@@ -784,6 +811,7 @@ class VideoPlayer extends Component {
             bufferForPlaybackAfterRebufferMs: 5000,
             cacheSizeMB: 200,
           }}
+          preventsDisplaySleepDuringVideoPlayback={true}
         />
       </TouchableOpacity>
     );