+++ ios/Video/RCTVideoManager.swift
@@ -11,77 +11,70 @@ class RCTVideoManager: RCTViewManager {
         return bridge.uiManager.methodQueue
     }
 
+    func performOnVideoView(withReactTag reactTag: NSNumber, callback: @escaping (RCTVideo?) -> Void) {
+        DispatchQueue.main.async { [weak self] in
+            guard let self else {
+                callback(nil)
+                return
+            }
+
+            guard let view = self.bridge.uiManager.view(forReactTag: reactTag) as? RCTVideo else {
+                RCTLogError("Invalid view returned from registry, expecting RCTVideo, got: \(String(describing: view))")
+                callback(nil)
+                return
+            }
+
+            callback(view)
+        }
+    }
+
     @objc(save:reactTag:resolver:rejecter:)
     func save(options: NSDictionary, reactTag: NSNumber, resolve: @escaping RCTPromiseResolveBlock, reject: @escaping RCTPromiseRejectBlock) {
-        bridge.uiManager.prependUIBlock { _, viewRegistry in
-            let view = viewRegistry?[reactTag]
-            if !(view is RCTVideo) {
-                RCTLogError("Invalid view returned from registry, expecting RCTVideo, got: %@", String(describing: view))
-            } else if let view = view as? RCTVideo {
-                view.save(options: options, resolve: resolve, reject: reject)
-            }
-        }
+        performOnVideoView(withReactTag: reactTag, callback: { videoView in
+            videoView?.save(options: options, resolve: resolve, reject: reject)
+        })
+    }
+
+    @objc(seek:reactTag:)
+    func seek(info: NSDictionary, reactTag: NSNumber) {
+        performOnVideoView(withReactTag: reactTag, callback: { videoView in
+            videoView?.setSeek(info)
+        })
     }
 
     @objc(setLicenseResult:licenseUrl:reactTag:)
     func setLicenseResult(license: NSString, licenseUrl: NSString, reactTag: NSNumber) {
-        bridge.uiManager.prependUIBlock { _, viewRegistry in
-            let view = viewRegistry?[reactTag]
-            if !(view is RCTVideo) {
-                RCTLogError("Invalid view returned from registry, expecting RCTVideo, got: %@", String(describing: view))
-            } else if let view = view as? RCTVideo {
-                view.setLicenseResult(license as String, licenseUrl as String)
-            }
-        }
+        performOnVideoView(withReactTag: reactTag, callback: { videoView in
+            videoView?.setLicenseResult(license as String, licenseUrl as String)
+        })
     }
 
     @objc(setLicenseResultError:licenseUrl:reactTag:)
     func setLicenseResultError(error: NSString, licenseUrl: NSString, reactTag: NSNumber) {
-        bridge.uiManager.prependUIBlock { _, viewRegistry in
-            let view = viewRegistry?[reactTag]
-            if !(view is RCTVideo) {
-                RCTLogError("Invalid view returned from registry, expecting RCTVideo, got: %@", String(describing: view))
-            } else if let view = view as? RCTVideo {
-                view.setLicenseResultError(error as String, licenseUrl as String)
-            }
-        }
+        performOnVideoView(withReactTag: reactTag, callback: { videoView in
+            videoView?.setLicenseResultError(error as String, licenseUrl as String)
+        })
     }
 
     @objc(dismissFullscreenPlayer:)
     func dismissFullscreenPlayer(_ reactTag: NSNumber) {
-        bridge.uiManager.prependUIBlock { _, viewRegistry in
-            let view = viewRegistry?[reactTag]
-            if !(view is RCTVideo) {
-                RCTLogError("Invalid view returned from registry, expecting RCTVideo, got: %@", String(describing: view))
-            } else if let view = view as? RCTVideo {
-                view.dismissFullscreenPlayer()
-            }
-        }
+        performOnVideoView(withReactTag: reactTag, callback: { videoView in
+            videoView?.dismissFullscreenPlayer()
+        })
     }
 
     @objc(presentFullscreenPlayer:)
     func presentFullscreenPlayer(_ reactTag: NSNumber) {
-        bridge.uiManager.prependUIBlock { _, viewRegistry in
-            let view = viewRegistry?[reactTag]
-            if !(view is RCTVideo) {
-                RCTLogError("Invalid view returned from registry, expecting RCTVideo, got: %@", String(describing: view))
-            } else if let view = view as? RCTVideo {
-                view.presentFullscreenPlayer()
-            }
-        }
+        performOnVideoView(withReactTag: reactTag, callback: { videoView in
+            videoView?.presentFullscreenPlayer()
+        })
     }
 
     @objc(setPlayerPauseState:reactTag:)
     func setPlayerPauseState(paused: NSNumber, reactTag: NSNumber) {
-        bridge.uiManager.prependUIBlock { _, viewRegistry in
-            let view = viewRegistry?[reactTag]
-            if !(view is RCTVideo) {
-                RCTLogError("Invalid view returned from registry, expecting RCTVideo, got: %@", String(describing: view))
-            } else if let view = view as? RCTVideo {
-                let paused = paused.boolValue
-                view.setPaused(paused)
-            }
-        }
+        performOnVideoView(withReactTag: reactTag, callback: { videoView in
+            videoView?.setPaused(paused.boolValue)
+        })
     }
 
     override class func requiresMainQueueSetup() -> Bool {